Adding L1SLOAD values in the end of L2 batches when writing to the DA layer
This improves the following two issues:
1. L2 nodes replaying transactions during validation no longer need an
archive node and don't need to call eth_getStorageAt for each L1SLOAD
they encounter in the batch, saving the extra RPC call for each of them,
at the expense of an extra 32 bytes written to DA, per call to the
L1SLOAD call in the batch. Removing the need for archive node and
improving validation speed seems worth it, especially when using blob
transactions.
2. L2 nodes don't need to worry about consistency when replaying
transaction during batch validation.